services/agentic-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.35.0
botocore==1.35.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
fastapi==0.115.14
frozenlist==1.7.0
h11==0.16.0
idna==3.10
lxml==6.0.0
mangum==0.19.0
multidict==6.6.3
numpy==2.3.1
openai==0.28.0
openpyxl==3.1.5
pandas==2.3.0
pandas-stubs==2.3.0.250703
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
types-pytz==2025.2.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
xlsxwriter==3.2.5
yarl==1.20.1
